I _think_ this fixes a bug by which WebKit on Mac was not highlighting newly added nodes.
Not sure what the bug was/is, but I have reason to believe it might have been due to an ajax race condition while reading the nodeIdAdded global variable.  Not sure why this variable was there, or what else its purpose is, so for now I've otherwise left it there.